Moved additional wrapper functions
Moved additional wrapper functions (RandomGen and TorusUtils) used only
in unittest to separate class in test module
