+++ libtfhe/core/arithmetic/torus_utils_big.h
@@ -6,7 +6,7 @@
 
 template<>
 class TorusUtils<BigTorus> {
-private:
+protected:
     using TORUS = BigTorus;
 
 public:
@@ -28,15 +28,6 @@ public:
     */
     static double to_double(const TORUS *x, const ZModuleParams<TORUS> *params) {
         return tfhe_backend::to_double(x, params);
-    }
-
-    /**
-     * @brief Converts torus to real number (mainly for printing)
-     * @param x torus element to convert
-     * @return real number
-    */
-    static double to_double(const TORUS &x, const ZModuleParams<TORUS> *params) {
-        return tfhe_backend::to_double(&x, params);
     }
 
     /**
@@ -82,17 +73,6 @@ public:
         double diff = std::abs(TorusUtils<TORUS>::to_double(t1, params)-TorusUtils<TORUS>::to_double(t2, params));
         diff = (diff > 0.5) ? 1.0 - diff : diff;
         return diff;
-    }
-
-    /**
-     * @brief Return infinity norm between 2 torus elements
-     *
-     * @param t1 first torus element
-     * @param t2 second torus element
-     * @return double value of the infinity norm
-     */
-    static double normInftyDist(const TORUS &t1, const TORUS &t2, const ZModuleParams<TORUS> *params) {
-        return normInftyDist(&t1, &t2, params);
     }
 };
